Refactor indexer
